Moving viewModel & test  to terms package
