+++ app/src/main/java/io/gnosis/safe/di/modules/ViewModelFactoryModule.kt
@@ -12,7 +12,7 @@ import io.gnosis.safe.ui.safe.balances.SafeBalancesViewModel
 import io.gnosis.safe.ui.safe.empty.NoSafeViewModel
 import io.gnosis.safe.ui.safe.selection.SafeSelectionViewModel
 import io.gnosis.safe.ui.safe.settings.SafeSettingsViewModel
-import io.gnosis.safe.ui.splash.SplashViewModel
+import io.gnosis.safe.ui.safe.terms.TermsViewModel
 import javax.inject.Singleton
 
 @Module
@@ -20,8 +20,8 @@ abstract class ViewModelFactoryModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(SplashViewModel::class)
-    abstract fun bindsSplashViewModel(viewModel: SplashViewModel): ViewModel
+    @ViewModelKey(TermsViewModel::class)
+    abstract fun bindsSplashViewModel(viewModel: TermsViewModel): ViewModel
 
     @Binds
     @IntoMap
