+++ app/src/main/res/layout/activity_splash.xml
@@ -12,9 +12,9 @@
         android:id="@+id/continue_button"
         style="@style/PrimaryButton"
         android:layout_width="0dp"
-        android:layout_marginEnd="16dp"
-        android:layout_marginStart="16dp"
-        android:text="Continue"
+        android:layout_marginEnd="@dimen/default_horizontal_margin"
+        android:layout_marginStart="@dimen/default_horizontal_margin"
+        android:text="Get started"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
