+++ app/src/test/java/io/gnosis/safe/ui/safe/splash/TermsViewModelTest.kt
@@ -2,7 +2,7 @@ package io.gnosis.safe.ui.safe.splash
 
 import android.app.Application
 import android.content.Context
-import io.gnosis.safe.ui.splash.SplashViewModel
+import io.gnosis.safe.ui.safe.terms.TermsViewModel
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.spyk
@@ -12,7 +12,7 @@ import org.junit.Test
 import pm.gnosis.svalinn.common.PreferencesManager
 import pm.gnosis.tests.utils.TestPreferences
 
-class SplashViewModelTest {
+class TermsViewModelTest {
 
     private lateinit var preferences: TestPreferences
     private lateinit var preferencesManager: PreferencesManager
@@ -29,9 +29,9 @@ class SplashViewModelTest {
 
     @Test
     fun `If terms already agreed, advance() method should be called and bottom sheet should not be shown`() {
-        preferences.edit().putBoolean(SplashViewModel.TERMS_AGREED, true).apply()
+        preferences.edit().putBoolean(TermsViewModel.TERMS_AGREED, true).apply()
 
-        val viewModel = SplashViewModel(preferencesManager)
+        val viewModel = TermsViewModel(preferencesManager)
         viewModel.termsBottomSheetDialog = mockk(relaxed = true)
 
         val callback: () -> Unit = mockk(relaxed = true)
@@ -44,9 +44,9 @@ class SplashViewModelTest {
 
     @Test
     fun `If terms not agreed, advance() method should not be called when agree button not hit`() {
-        preferences.edit().putBoolean(SplashViewModel.TERMS_AGREED, false).apply()
+        preferences.edit().putBoolean(TermsViewModel.TERMS_AGREED, false).apply()
 
-        val viewModel = SplashViewModel(preferencesManager)
+        val viewModel = TermsViewModel(preferencesManager)
         viewModel.termsBottomSheetDialog = mockk(relaxed = true)
         val callback: () -> Unit = mockk(relaxed = true)
